refactor: 프론트 기능/컴포넌트 파일/폴더 분리 (#45)
