+++ docs/dev-logs/2026-04-22-frontend-file-folder-split.md
@@ -1,0 +1,39 @@
+# 2026-04-22 Frontend File/Folder Split
+
+## 대상
+
+- 브랜치: `feat/45-slice`
+- 범위: 이슈 `#45` 프론트 기능/컴포넌트 단위 파일/폴더 분리
+
+## 기준
+
+- `App.tsx` 오케스트레이션 축소
+- `views`/`features`/`shared` 책임 분리
+- 기존 상호작용/접근성 유지
+
+## A/B 비교
+
+### A. App 중심 유지 + 일부 유틸만 추출
+
+- 장점: 작업이 단순하다.
+- 단점: 구조 개선 효과가 제한적이다.
+
+### B. 화면/기능/공통 계층으로 전면 분리
+
+- 장점: 변경 영향 범위가 작아지고 재사용성이 좋아진다.
+- 단점: 파일 이동/연결 검증 비용이 증가한다.
+
+## 선택
+
+- `B`를 선택했다.
+- 이유: 이슈의 수용 기준(단일 대형 파일 의존도 감소, 폴더 책임 명확화)을 충족한다.
+
+## 검증
+
+- `npm run lint` 통과
+- `npm run build` 통과
+
+## 리스크
+
+- 현재 ESLint 설정에서 TS 타입 시그니처가 `no-unused-vars`와 충돌하는 지점이 있어 일부 파일 단위 완화가 포함됐다.
+- 후속으로 ESLint 규칙을 `@typescript-eslint/no-unused-vars` 기준으로 정렬하는 작업이 필요하다.
