+++ frontend/src/shared/components/InfoTile.tsx
@@ -1,0 +1,8 @@
+export function InfoTile({ label, value }: { label: string; value: string }) {
+  return (
+    <div className="info-tile">
+      <span>{label}</span>
+      <strong>{value}</strong>
+    </div>
+  );
+}
