+++ frontend/src/shared/components/Panel.tsx
